This single patch contains all the WebAudio related commits in WebKit
trunk from May 15, 2011 to Feb 1, 2012

This patch contains no integration or build fixes and WebAudio feature
is not enabled in this patch.

There are no build breakes or regressions in WebKit because of this
patch.

Cherry-picked from change-Id: I49c3905497931bc9993ae40869a9bb30a2dfa7e3

(cherry picked from commit cd2cfa5e9db3316ee09a939b2ff2c76daa3fb3a0)

+++ Source/WebCore/platform/audio/mac/AudioFileReaderMac.h
@@ -36,7 +36,7 @@
 namespace WebCore {
 
 class AudioBus;
-    
+
 // Wrapper class for AudioFile and ExtAudioFile CoreAudio APIs for reading files and in-memory versions of them...
 
 class AudioFileReader {
@@ -46,7 +46,7 @@ public:
     ~AudioFileReader();
 
     // Returns 0 if error
-    PassOwnPtr<AudioBus> createBus(double sampleRate, bool mixToMono);
+    PassOwnPtr<AudioBus> createBus(float sampleRate, bool mixToMono);
 
     const void* data() const { return m_data; }
     size_t dataSize() const { return m_dataSize; }
@@ -67,5 +67,5 @@ private:
 };
 
 } // namespace WebCore
- 
+
 #endif // AudioFileReaderMac_h
