+++ Source/WebCore/webaudio/AudioBuffer.h
@@ -29,7 +29,7 @@
 #ifndef AudioBuffer_h
 #define AudioBuffer_h
 
-#include "Float32Array.h"
+#include <wtf/Float32Array.h>
 #include <wtf/PassRefPtr.h>
 #include <wtf/RefCounted.h>
 #include <wtf/RefPtr.h>
@@ -38,18 +38,18 @@
 namespace WebCore {
 
 class AudioBus;
-    
+
 class AudioBuffer : public RefCounted<AudioBuffer> {
-public:   
-    static PassRefPtr<AudioBuffer> create(unsigned numberOfChannels, size_t numberOfFrames, double sampleRate);
+public:
+    static PassRefPtr<AudioBuffer> create(unsigned numberOfChannels, size_t numberOfFrames, float sampleRate);
 
     // Returns 0 if data is not a valid audio file.
-    static PassRefPtr<AudioBuffer> createFromAudioFileData(const void* data, size_t dataSize, bool mixToMono, double sampleRate);
+    static PassRefPtr<AudioBuffer> createFromAudioFileData(const void* data, size_t dataSize, bool mixToMono, float sampleRate);
 
     // Format
     size_t length() const { return m_length; }
     double duration() const { return length() / sampleRate(); }
-    double sampleRate() const { return m_sampleRate; }
+    float sampleRate() const { return m_sampleRate; }
 
     // Channel data access
     unsigned numberOfChannels() const { return m_channels.size(); }
@@ -64,13 +64,13 @@ public:
     // releaseMemory() can be called when the AudioContext goes away, so we can release the memory earlier than when the garbage collection happens.
     // Careful! Only call this when the page unloads, after the AudioContext is no longer processing.
     void releaseMemory();
-    
+
 protected:
-    AudioBuffer(unsigned numberOfChannels, size_t numberOfFrames, double sampleRate);
+    AudioBuffer(unsigned numberOfChannels, size_t numberOfFrames, float sampleRate);
     AudioBuffer(AudioBus* bus);
 
     double m_gain; // scalar gain
-    double m_sampleRate;
+    float m_sampleRate;
     size_t m_length;
 
     Vector<RefPtr<Float32Array> > m_channels;
