+++ Source/WebCore/webaudio/HighPass2FilterNode.h
@@ -31,19 +31,19 @@
 namespace WebCore {
 
 class AudioParam;
-    
+
 class HighPass2FilterNode : public AudioBasicProcessorNode {
 public:
-    static PassRefPtr<HighPass2FilterNode> create(AudioContext* context, double sampleRate)
+    static PassRefPtr<HighPass2FilterNode> create(AudioContext* context, float sampleRate)
     {
-        return adoptRef(new HighPass2FilterNode(context, sampleRate));      
+        return adoptRef(new HighPass2FilterNode(context, sampleRate));
     }
 
     AudioParam* cutoff() { return biquadProcessor()->parameter1(); }
     AudioParam* resonance() { return biquadProcessor()->parameter2(); }
-    
+
 private:
-    HighPass2FilterNode(AudioContext*, double sampleRate);
+    HighPass2FilterNode(AudioContext*, float sampleRate);
 
     BiquadProcessor* biquadProcessor() { return static_cast<BiquadProcessor*>(processor()); }
 };
